+++ Streamlit_Datamon/ITERATION_LOG.MD
@@ -1,0 +1,12 @@
+ME: I pasted the code to our memorybank before this message: i am a junior developer learning the python streamlit library, you are an expert in streamlit. please help me convert this working program to a streamlit application and give me a simplified explanation of what you did as well. thank you
+ChatGPT gave me a mostly functioning application, with some design issues that i was able to change with follow up prompts. Program ended up working with many tests.
+
+ME: I pasted the code to our number-guesser before this message: please turn this cli app into a streamlit application, keeping the current logic of the program
+ChatGPT provided a working application that had some problems with the buttons. Through a few more prompts the buttons were improved to make the game better for the user.
+
+ME: I pasted the code to our memorybank before this message: i am continuing to learn streamlit, and would like you to convert this program into another streamlit application
+ChatGPT gave me a working program but I changed the box used to enter the number from a number input that had arrows on the side to a standard text input and changed it to an int.
+
+ME: I pasted the code to all of the games before this message:  can you show me how i would take these 4 seperate games and make them in one application with a homepage, where you would choose which game to go play and can return to from the game
+ChatGPT gave me a propgram that would create a homepage and some code to insert into all other files, that allowed the main file to call each game and be able to return to the homepage. Needs more work, but the games 
+can be played.
