+++ Streamlit_Datamon/README.MD
@@ -1,0 +1,3 @@
+To start the application run these commands in the command line
+pip install streamlit (if not installed already)
+streamlit run datamon_complete.py
