+++ Streamlit_Datamon/answerchecker.py
@@ -1,88 +1,96 @@
 import streamlit as st
 import operator
 
-# Initialize session state variables to persist values between interactions
-if "score" not in st.session_state:
-    st.session_state.score = 0
-if "total_questions" not in st.session_state:
-    st.session_state.total_questions = 0
-if "game_over" not in st.session_state:
-    st.session_state.game_over = False
-# keep these in session state for compatibility, but they are not required for the new behavior
-if "show_correct" not in st.session_state:
-    st.session_state.show_correct = False
-if "correct_to_show" not in st.session_state:
-    st.session_state.correct_to_show = None
+def run_game():
+    st.title("🧠 Answer Checker")
 
-st.title("🧠 Equation Answer Checker")
-st.write("Enter a basic equation (e.g., `3 + 5 = 8`) and see if you got it right! Supports +, -, *, and / (integer division).")
+    # Back button
+    if st.button("⬅️ Back to Main Menu"):
+        st.session_state.page = "home"
+        st.rerun()
 
-# Quit button only
-if st.button("Quit Game"):
-    st.session_state.game_over = True
-
-if not st.session_state.game_over:
-    # Equation input field
-    equation = st.text_input("Enter your equation here:", placeholder="e.g. 7 * 6 = 42")
-
-    if st.button("Check Answer"):
-        # Reset any previous "show correct" state
+    # Initialize session state variables to persist values between interactions
+    if "score" not in st.session_state:
+        st.session_state.score = 0
+    if "total_questions" not in st.session_state:
+        st.session_state.total_questions = 0
+    if "game_over" not in st.session_state:
+        st.session_state.game_over = False
+    # keep these in session state for compatibility, but they are not required for the new behavior
+    if "show_correct" not in st.session_state:
         st.session_state.show_correct = False
+    if "correct_to_show" not in st.session_state:
         st.session_state.correct_to_show = None
 
-        if "=" not in equation:
-            st.warning("❌ Invalid format. Make sure to include `=` in your equation.")
-        else:
-            expr, userAns = equation.split("=", 1)
-            expr = expr.strip()
-            userAns = userAns.strip()
+    st.title("🧠 Equation Answer Checker")
+    st.write("Enter a basic equation (e.g., `3 + 5 = 8`) and see if you got it right! Supports +, -, *, and / (integer division).")
 
-            # Detect the operator
-            try:
-                operation = next(op for op in ['+', '-', '*', '/'] if op in expr)
-            except StopIteration:
-                st.error("Invalid equation. Please include +, -, *,, or /.")
-                st.stop()
+    # Quit button only
+    if st.button("Quit Game"):
+        st.session_state.game_over = True
 
-            # Split into numbers
-            try:
-                num1, num2 = expr.split(operation)
-                num1 = int(num1.strip())
-                num2 = int(num2.strip())
-                userAns = int(userAns)
-            except Exception:
-                st.error("Invalid equation, only two numbers and one operator. Try again.")
-                st.stop()
+    if not st.session_state.game_over:
+        # Equation input field
+        equation = st.text_input("Enter your equation here:", placeholder="e.g. 7 * 6 = 42")
 
-            # Evaluate the answer
-            operatorsDict = {
-                '+': operator.add(num1, num2),
-                '-': operator.sub(num1, num2),
-                '*': operator.mul(num1, num2),
-                '/': operator.floordiv(num1, num2),
-            }
-            correct_ans = operatorsDict[operation]
+        if st.button("Check Answer"):
+            # Reset any previous "show correct" state
+            st.session_state.show_correct = False
+            st.session_state.correct_to_show = None
 
-            # Check correctness
-            if userAns == correct_ans:
-                st.success("✅ You got it right! Good job!")
-                st.session_state.score += 1
+            if "=" not in equation:
+                st.warning("❌ Invalid format. Make sure to include `=` in your equation.")
             else:
-                st.error("❌ Your answer is incorrect.")
-                # Store the correct answer so it persists if needed elsewhere
-                st.session_state.correct_to_show = correct_ans
-                # Immediately show the correct answer in a compact info box
-                st.info(f"Correct answer: {st.session_state.correct_to_show}")
+                expr, userAns = equation.split("=", 1)
+                expr = expr.strip()
+                userAns = userAns.strip()
 
-            st.session_state.total_questions += 1
+                # Detect the operator
+                try:
+                    operation = next(op for op in ['+', '-', '*', '/'] if op in expr)
+                except StopIteration:
+                    st.error("Invalid equation. Please include +, -, *,, or /.")
+                    st.stop()
 
-    # Display score
-    st.write(f"**Score:** {st.session_state.score} out of {st.session_state.total_questions}")
+                # Split into numbers
+                try:
+                    num1, num2 = expr.split(operation)
+                    num1 = int(num1.strip())
+                    num2 = int(num2.strip())
+                    userAns = int(userAns)
+                except Exception:
+                    st.error("Invalid equation, only two numbers and one operator. Try again.")
+                    st.stop()
 
-else:
-    st.subheader("🎯 Final Score")
-    st.write(f"You scored **{st.session_state.score}** out of **{st.session_state.total_questions}** questions.")
-    if st.button("Restart Game"):
-        st.session_state.score = 0
-        st.session_state.total_questions = 0
-        st.session_state.game_over = False
+                # Evaluate the answer
+                operatorsDict = {
+                    '+': operator.add(num1, num2),
+                    '-': operator.sub(num1, num2),
+                    '*': operator.mul(num1, num2),
+                    '/': operator.floordiv(num1, num2),
+                }
+                correct_ans = operatorsDict[operation]
+
+                # Check correctness
+                if userAns == correct_ans:
+                    st.success("✅ You got it right! Good job!")
+                    st.session_state.score += 1
+                else:
+                    st.error("❌ Your answer is incorrect.")
+                    # Store the correct answer so it persists if needed elsewhere
+                    st.session_state.correct_to_show = correct_ans
+                    # Immediately show the correct answer in a compact info box
+                    st.info(f"Correct answer: {st.session_state.correct_to_show}")
+
+                st.session_state.total_questions += 1
+
+        # Display score
+        st.write(f"**Score:** {st.session_state.score} out of {st.session_state.total_questions}")
+
+    else:
+        st.subheader("🎯 Final Score")
+        st.write(f"You scored **{st.session_state.score}** out of **{st.session_state.total_questions}** questions.")
+        if st.button("Restart Game"):
+            st.session_state.score = 0
+            st.session_state.total_questions = 0
+            st.session_state.game_over = False
